requirements.txt
Django==1.10
djangorestframework==3.7.7
jsonfield==2.0.2